pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["BangPham <bang.pham@paradox.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.1.1"
djangorestframework = "^3.14.0"
python-dotenv = "^0.21.0"
djangorestframework-simplejwt = "^5.2.0"
drf-yasg = "^1.21.4"
django-cors-headers = "^3.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"